refactor(mdast): extract JSON scalar helpers into escape module
Splits mdast.rs (397 lines) under the 350-line source-file limit by
moving the string-escaping scan, integer writer, escape-flag table, and
byte-escape helper into a sibling mdast/escape module. No behavior
change.
